refactor: use @elgato/schemas for manifest and layout definitions
Greatly improve the types exported by @elgato/streamdeck.
