+++ src/api/index.ts
@@ -1,4 +1,4 @@
-import { Manifest as ManifestSchema } from "@elgato/schemas/streamdeck/plugins";
+import type * as schemas from "@elgato/schemas/streamdeck/plugins";
 
 export * from "./command";
 export * from "./device";
@@ -11,4 +11,4 @@ export * from "./target";
 /**
  * Defines the plugin and available actions, and all information associated with them, including the plugin's entry point, all iconography, action default behavior, etc.
  */
-export type Manifest = Omit<ManifestSchema, "$schema">;
+export type Manifest = Omit<schemas.Manifest, "$schema">;
