+++ src/plugin/__mocks__/manifest.ts
@@ -33,6 +33,7 @@ export const manifest: Manifest = {
 	Software: {
 		MinimumVersion: "6.5"
 	},
+	UUID: "com.elgato.test",
 	Version: "1.0.0"
 };
