+++ src/plugin/index.ts
@@ -1,8 +1,25 @@
 import { StreamDeck } from "./stream-deck";
 
-export { DeviceType, Target, type Language, type Manifest, type PayloadObject, type RegistrationInfo } from "../api";
-export * from "../api/layout";
-export { Action } from "./actions/action";
+export {
+	Bar,
+	BarSubType,
+	Controller,
+	Coordinates,
+	DeviceInfo,
+	DeviceType,
+	FeedbackPayload,
+	GBar,
+	Language,
+	Manifest,
+	PayloadObject,
+	Pixmap,
+	RegistrationInfo,
+	Size,
+	State,
+	Target,
+	Text
+} from "../api";
+export { Action, ImageOptions, TitleOptions, TriggerDescriptionOptions } from "./actions/action";
 export { action } from "./actions/decorators";
 export { SingletonAction } from "./actions/singleton-action";
 export { EventEmitter, EventsOf } from "./common/event-emitter";
